+++ perl/FSM/Scheduler/ISF/LoweringIR.pm
@@ -328,6 +328,7 @@ sub _build_transaction($self, $tx, $actor, $txi) {
     my $tn  = $tx->{name};
     my $wd  = $actor->{watchdog};
     my $drives = $actor->{drives} || {};
+    _reject_unsupported_contract_clauses($tx->{clauses}, $tn);
     my $widths = _build_signal_width_map($actor, $tx);
     my @st; my %ct; my @dt; my @ps; my @doc; my @spc; my @dps;
     my $si  = 0; my $ha = 0; my $wdc; my $lat;
@@ -394,6 +395,28 @@ sub _build_transaction($self, $tx, $actor, $txi) {
     $ct{can_accept} = 1;
     for my $s (@st) { next unless $s->{kind} eq 'entry'; unshift @{$s->{assignments}}, { lhs => 'can_accept', rhs => 1, op => '=' }; }
     return (\@st, \%ct, \@dt, \@doc, \@spc);
+}
+
+sub _reject_unsupported_contract_clauses {
+    my ($clauses, $tn) = @_;
+    return unless ref($clauses) eq 'ARRAY';
+
+    for my $clause (@$clauses) {
+        next unless ref($clause) eq 'ARRAY' && @$clause;
+        my $keyword = $clause->[0];
+        if (defined($keyword) && !ref($keyword) && $keyword eq 'contract') {
+            confess "Transaction '$tn': temporal '(contract ...)' clauses are parsed but not implemented by ISF lowering\n";
+        }
+
+        if (defined($keyword) && !ref($keyword) && ($keyword eq 'when' || $keyword eq 'repeat')) {
+            _reject_unsupported_contract_clauses([@{$clause}[2 .. $#$clause]], $tn);
+        } elsif (defined($keyword) && !ref($keyword) && $keyword eq 'switch') {
+            for my $branch (@{$clause}[2 .. $#$clause]) {
+                next unless ref($branch) eq 'ARRAY';
+                _reject_unsupported_contract_clauses([@{$branch}[1 .. $#$branch]], $tn);
+            }
+        }
+    }
 }
 
 # --- Individual clause → IR ---